Fixed up a few errors
When checking for equality you have to use == not just =. It also wasn't clear how to select the dice. I thought I would use D2 instead of the number 1. So I kind of fixed that up a bit. I didn't finalize everything you will need to still address some of the issues with the global variable "dice" and how to get the selection for the dice so that the results can be appropriate to what the user selected.